Fix stability and security problem (#117)
* Keeping reference to parser pool

This change ensure that the parser is always returned to the pool which it was taken from. It was not guarantied that when the finally-block called parserPool.get() that it would get the same pool from the ContextClassloaderLocal which the parser was acquired from, because they are cached in a WeakHashMap, which may have been GC-ed between the time the parser was taken, and later returned. The change keeps the reference to the ParserPool in scope to also be used to return the parser.

Bug: https://github.com/eclipse-ee4j/metro-saaj/issues/73#issuecomment-424014253
Signed-off-by: Stefan Magnus Landrø <stefan.landro@gmail.com>
Signed-off-by: Rune Flobakk <flobakk@gmail.com>

* Configurable size limit for the SOAP body part

Can set saaj.mime.soapBodyPartSizeLimit system property to configure max allowed size of the soapMessagePart to prevent maliciously crafted messages to take down the server with OutOfMemoryError.

Bug: https://github.com/eclipse-ee4j/metro-saaj/issues/77
Signed-off-by: Stefan Magnus Landrø <stefan.landro@gmail.com>
Signed-off-by: Rune Flobakk <flobakk@gmail.com>