Bug 165 - Document owner is inconsistent
This change fixes the delegation so that element types, including attributes and text nodes, correctly identify as being part of the same owning document.
Note that whilst this fixes the consistency of some cloned elements, a proper fix is beyond what can be achieved within this change and requires a
comprehensive review.

Signed-off-by: Mark Brodziak <mark.brodziak+eclipse@gmail.com>